LPS-56013 Move Basic and Parameter AuthVerifiers into own package for consistency
