+++ portal-impl/src/portal.properties
@@ -3611,7 +3611,7 @@
     # This property is not read by the portal except for portal properties
     # overridden by liferay-hook.xml. It remains here only as a reference.
     #
-    #auth.verifier.pipeline=com.liferay.portal.security.auth.verifier.BasicAuthHeaderAuthVerifier,com.liferay.portal.security.auth.verifier.DigestAuthenticationAuthVerifier,com.liferay.portal.security.auth.verifier.ParameterAuthVerifier,com.liferay.portal.security.auth.verifier.PortalSessionAuthVerifier,com.liferay.portal.security.auth.verifier.TunnelingServletAuthVerifier
+    #auth.verifier.pipeline=com.liferay.portal.security.auth.verifier.basic.BasicAuthHeaderAuthVerifier,com.liferay.portal.security.auth.verifier.DigestAuthenticationAuthVerifier,com.liferay.portal.security.auth.verifier.parameter.ParameterAuthVerifier,com.liferay.portal.security.auth.verifier.PortalSessionAuthVerifier,com.liferay.portal.security.auth.verifier.TunnelingServletAuthVerifier
 
     #
     # Each authentication verifier can have custom properties set via the
